requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.3
django-environ==0.4.5
django-model-utils==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
entrypoints==0.3
factory-boy==2.12.0
Faker==1.0.7
flake8==3.7.8
importlib-metadata==0.18
mccabe==0.6.1
more-itertools==7.1.0
mypy==0.711
mypy-extensions==0.4.1
packaging==19.0
pendulum==2.0.5
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.1
pytzdata==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.0
typed-ast==1.4.0
wcwidth==0.1.7
zipp==0.5.2
